Refactor authentication logic: update token expiration times, rename verifyAcsessToken to getProfile, and move verifyToken utility to a new location
