+++ src/routes/auth.routes.ts
@@ -1,5 +1,6 @@
 import { Router } from 'express';
-import { refreshToken, signIn, signOut, signUp, verifyAcsessToken, verifyEmail } from '../controllers/auth.controller';
+import { refreshToken, signIn, signOut, signUp, getProfile, verifyEmail } from '../controllers/auth.controller';
+import { authorize } from '../middlewares/auth.middleware';
 
 const authRouter = Router();
 
@@ -8,6 +9,6 @@ authRouter.post('/sign-in', signIn);
 authRouter.post("/sign-out", signOut);
 authRouter.post('/refresh-token', refreshToken);
 authRouter.post('/verify-email', verifyEmail);
-authRouter.post('/verify-acsess-token', verifyAcsessToken);
+authRouter.post('/verify-acsess-token', authorize, getProfile);
 
 export default authRouter;
