--- src/utils/errors/auth/http.auth.ts
@@ -1,12 +1,0 @@
-import jwt from "jsonwebtoken";
-
-export function verifyToken(
-  token: string,
-  jwt_secret: string
-): { id: string } | null {
-  try {
-    return jwt.verify(token, jwt_secret) as { id: string };
-  } catch {
-    return null;
-  }
-}
